Fix apparent use-after-free in nvrtc.
It seems like in recent nvrtc versions (at least in 12.8),
the program source code and name are not copied in nvrtcCreateProgram,
but instead kept as a reference. Previously, cudarc would create
a temporary CString which was dropped before destroying the nvrtcProgram.
This led to various miscompilations and crashes in nvrtc.
This commit fixes the problem by keeping the corresponding CStrings
alive for the lifetime of the nvrtcProgram.